pyproject.toml
[tool.poetry]
name = "hevyai"
version = "0.1.0"
description = "Aplicação para consumir a API do Hevy utilizando Django, DRF, React e TailwindCSS"
authors = ["Seu Nome <seu.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
whitenoise = "^6.5.0"
django-cors-headers = "^4.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-django = "^4.5.2"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88